Move all single-method plotting wrappers to Axes
